[NTOS:KE] Preserve 64-bit affinity when selecting process thread seeds
KeFindNextRightSetAffinity accepts KAFFINITY. Avoid truncating CPUs 32-63 on amd64.

+++ ntoskrnl/ke/procobj.c
@@ -196,7 +196,7 @@ KeInitializeProcess(IN OUT PKPROCESS Process,
     /* Find the matching affinity set to calculate the thread seed */
     Affinity &= Node->ProcessorMask;
     Process->ThreadSeed = KeFindNextRightSetAffinity(Node->Seed,
-                                                     (ULONG)Affinity);
+                                                     Affinity);
     Node->Seed = Process->ThreadSeed;
 #endif
 }
